+++ src/components/login/Login.tsx
@@ -1,0 +1,7 @@
+import './Login.css'
+
+export const Login = () => (
+  <div>
+    <h1>Pantalla de login</h1>
+  </div>
+)
